+++ variables.tf
@@ -111,8 +111,14 @@ variable "podnet_cidr_block" {
   default     = "100.65.0.0/16"
 }
 
+# variable "nat_gateway_id" {
+#   description = "ID of a user-assigned NAT Gateway to use for cluster egress traffic, if not set a cluster managed load balancer will be used."
+#   type        = string
+#   default     = null
+# }
+
 variable "managed_outbound_ip_count" {
-  description = "Count of desired managed outbound IPs for the cluster load balancer. Must be between 1 and 100 inclusive."
+  description = "Count of desired managed outbound IPs for the cluster managed load balancer. Ignored if NAT gateway is specified, must be between 1 and 100 inclusive."
   type        = number
   default     = 1
 
@@ -123,7 +129,7 @@ variable "managed_outbound_ip_count" {
 }
 
 variable "managed_outbound_ports_allocated" {
-  description = "Number of desired SNAT port for each VM in the clusters load balancer. Must be between 0 & 64000 inclusive and divisible by 8."
+  description = "Number of desired SNAT port for each VM in the cluster managed load balancer. Ignored if NAT gateway is specified, must be between 0 & 64000 inclusive and divisible by 8."
   type        = number
   default     = 0
 
@@ -134,7 +140,7 @@ variable "managed_outbound_ports_allocated" {
 }
 
 variable "managed_outbound_idle_timeout" {
-  description = "Desired outbound flow idle timeout in seconds for the cluster load balancer. Must be between 240 and 7200 inclusive."
+  description = "Desired outbound flow idle timeout in seconds for the cluster managed load balancer. Ignored if NAT gateway is specified, must be between 240 and 7200 inclusive."
   type        = number
   default     = 240